Revert "IN-129 restructuring for clarity (#35)"
This reverts commit c4426876a02e98f15684643c16cf1f19f1f9fe13.
