+++ lambda_functions/v1/functions/supporting_docs/supporting_docs.py
@@ -1,4 +1,5 @@
 import datetime
+import json
 import logging
 import os
 from urllib.parse import urljoin, urlparse
@@ -8,11 +9,7 @@ import jwt
 import requests
 from botocore.exceptions import ClientError
 
-
-# Sirius API Service
-
 logger = logging.getLogger()
-
 try:
     logger.setLevel(os.environ["LOGGER_LEVEL"])
 except KeyError:
@@ -23,6 +20,131 @@ handler.setFormatter(
     logging.Formatter("[%(levelname)s] [in %(funcName)s:%(lineno)d] %(message)s")
 )
 logger.addHandler(handler)
+
+
+def lambda_handler(event, context):
+    """
+
+    Args:
+        event: json received from API Gateway
+        context:
+    Returns:
+        Response from Sirius in AWS Lambda format, json
+    """
+
+    sirius_api_url = build_sirius_url(
+        base_url=os.environ["SIRIUS_BASE_URL"],
+        api_route=os.environ["SIRIUS_PUBLIC_API_URL"],
+        endpoint="documents",
+    )
+    sirius_payload = transform_event_to_sirius_request(event=event)
+    sirius_headers = build_sirius_headers()
+
+    lambda_response = submit_document_to_sirius(
+        url=sirius_api_url, data=sirius_payload, headers=sirius_headers
+    )
+    logger.debug(f"Lambda Response: {lambda_response}")
+
+    return lambda_response
+
+
+def validate_event(event):
+    # TODO if there is not a nicer way to do this, there should be
+    """
+    The request body *should* be validated by API-G before it gets this far,
+    but given everything blows up if any of these required fields are missing/wrong
+    then it's worth double checking here, and providing integrators with a meaningful
+    error message
+
+    Args:
+        event: AWS event json
+
+    Returns:
+        tuple: valid boolean, error list
+    """
+
+    required_body_structure = {
+        "supporting_document": {
+            "data": {
+                "attributes": {"submission_id": 0},
+                "file": {"name": "string", "mimetype": "string", "source": "string"},
+            }
+        }
+    }
+
+    errors = compare_two_dicts(
+        required_body_structure, json.loads(event["body"]), missing=[]
+    )
+
+    if len(errors) > 0:
+        return False, errors
+    else:
+        return True, errors
+
+
+def transform_event_to_sirius_request(event):
+    """
+    Takes the 'body' from the AWS event and converts it into the right format for the
+    Sirius documents endpoint, detailed here:
+    tests/test_data/sirius_documents_payload_schema.json
+
+    Args:
+        event: json received from API Gateway
+    Returns:
+        Sirius-style payload, json
+    """
+    report_id = event["pathParameters"]["id"]
+    case_ref = event["pathParameters"]["caseref"]
+    request_body = json.loads(event["body"])
+    metadata = request_body["supporting_document"]["data"]["attributes"]
+    metadata["report_id"] = report_id
+    file_name = request_body["supporting_document"]["data"]["file"]["name"]
+    file_type = request_body["supporting_document"]["data"]["file"]["mimetype"]
+    file_source = request_body["supporting_document"]["data"]["file"]["source"]
+
+    payload = {
+        "type": "Report",
+        "caseRecNumber": case_ref,
+        "metadata": metadata,
+        "file": {"name": file_name, "source": file_source, "type": file_type},
+    }
+    logger.debug(f"Sirius Payload: {payload}")
+
+    return json.dumps(payload)
+
+
+# Helpers
+
+
+def compare_two_dicts(required_structure, test_dict, path="", missing=[]):
+
+    for key in required_structure:
+        if key not in test_dict:
+            missing_item = f"{path}->{key}"
+            if missing_item not in missing:
+                missing.append(missing_item)
+        else:
+            if type(required_structure[key]) is dict:
+                if path == "":
+                    path = key
+                else:
+                    path = path + "->" + key
+                compare_two_dicts(
+                    required_structure[key], test_dict[key], path, missing
+                )
+            else:
+                if isinstance(test_dict[key], type(None)):
+                    missing.append(f"{path}->{key}")
+                elif type(test_dict[key]) == str and len(test_dict[key]) == 0:
+                    missing_item = f"{path}->{key}"
+
+                    if missing_item not in missing:
+                        missing.append(missing_item)
+
+    return missing
+
+
+# Sirius API Service
 
 
 def build_sirius_url(base_url, api_route, endpoint):
